requirements.txt
asyncio 
flask 
dash 
plotly
pyserial
matplotlib 
numpy
pandas
scipy
dash
setuptools
dash-bootstrap-components
